web/api/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
apprise==1.9.6
asgiref==3.11.0
attrs==25.4.0
beautifulsoup4==4.12.2
billiard==4.2.4
blinker==1.9.0
cachetools==6.2.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.4
dataclasses-json==0.6.7
distro==1.9.0
Django==6.0.1
django-cors-headers==4.9.0
django-environ==0.12.0
django-fernet-fields==0.6
djangorestframework==3.16.1
fastapi==0.128.0
Flask==3.1.2
flask-cors==6.0.2
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
google-genai==1.55.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.0
groq==1.0.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.6.2
langchain==1.2.4
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
Markdown==3.10
MarkupSafe==3.0.3
marshmallow==3.26.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.4.1
oauthlib==3.3.1
openai==2.15.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==3.0.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.5
PyYAML==6.0.3
redis==7.1.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
sqlparse==0.5.5
starlette==0.50.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tqdm==4.67.1
twilio==9.9.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
uuid_utils==0.13.0
uvicorn==0.40.0
vine==5.1.0
wcwidth==0.5.0
websockets==15.0.1
Werkzeug==3.1.5
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
